Loading and Unloading of images
Re-separate the image/image_path getters and setters, and let user decide whether to "load" the image into the frame. The idea is that for lightweight calculations, the Frame will read the image from the path directly; but the user can keep the image loaded as a variable if it needs to be used multiple times (i.e. in EndtoEnd).